(PC-21560) refactor(InstalledMessagingApps): Create new component to simplify logic
