+++ src/features/offer/components/shareMessagingOffer/InstalledMessagingApps.tsx
@@ -1,16 +1,11 @@
 import React, { useState, useEffect } from 'react'
-import { Linking, Platform } from 'react-native'
-import Share, { ShareSingleOptions, Social } from 'react-native-share'
 
-import { MessagingAppContainer } from 'features/offer/components/shareMessagingOffer/MessagingAppContainer'
+import { MessagingAppButton } from 'features/offer/components/shareMessagingOffer/MessagingAppButton'
 import { getInstalledApps } from 'features/offer/helpers/getInstalledApps/getInstalledApps'
 import { getOfferUrl } from 'features/share/helpers/getOfferUrl'
 import { useShareOfferMessage } from 'features/share/helpers/useShareOfferMessage'
-import { analytics } from 'libs/analytics'
 import { eventMonitoring } from 'libs/monitoring'
-import { Network, ShareMessagingApp } from 'ui/components/ShareMessagingApp'
-
-const isWeb = Platform.OS === 'web'
+import { Network } from 'ui/components/ShareMessagingApp'
 
 export const InstalledMessagingApps = ({ offerId }: { offerId: number }) => {
   const [installedApps, setInstalledApps] = useState<Network[]>([])
@@ -27,81 +22,15 @@ export const InstalledMessagingApps = ({ offerId }: { offerId: number }) => {
 
   return (
     <React.Fragment>
-      {installedApps.map((network) => {
-        const {
-          shouldEncodeURI,
-          supportsURL = true,
-          isNative,
-          webUrl,
-          ...options
-        } = mapNetworkToSocial[network]
-        // Message has to be concatenated with url if url option is not supported, and in web
-        const message = supportsURL && !isWeb ? shareMessage : shareMessage + '\n' + shareUrl
-
-        const onPress = async () => {
-          analytics.logShare({ type: 'Offer', id: offerId, from: 'offer', social: options.social })
-          if (isWeb && webUrl) await Linking.openURL(webUrl + encodeURIComponent(message))
-          else if (isNative && options.url) await Linking.openURL(options.url + message)
-          else {
-            await Share.shareSingle({
-              message: shouldEncodeURI ? encodeURI(message) : message,
-              url: supportsURL ? shareUrl : undefined,
-              ...options,
-            })
-          }
-        }
-
-        return (
-          <MessagingAppContainer key={network}>
-            <ShareMessagingApp network={network} onPress={onPress} />
-          </MessagingAppContainer>
-        )
-      })}
+      {installedApps.map((network) => (
+        <MessagingAppButton
+          key={network}
+          offerId={offerId}
+          network={network}
+          shareMessage={shareMessage}
+          shareUrl={shareUrl}
+        />
+      ))}
     </React.Fragment>
   )
-}
-
-const mapNetworkToSocial: Record<
-  Network,
-  ShareSingleOptions & {
-    shouldEncodeURI?: boolean
-    supportsURL?: boolean
-    isNative?: boolean
-    webUrl?: string
-  }
-> = {
-  [Network.instagram]: {
-    social: Social.Instagram,
-    supportsURL: false,
-    shouldEncodeURI: Platform.OS === 'ios',
-    type: 'text',
-  },
-  [Network.messenger]: { social: Social.Messenger },
-  [Network.snapchat]: { social: Social.Snapchat },
-  [Network.googleMessages]: {
-    social: Social.Sms,
-    recipient: '',
-  },
-  [Network.whatsapp]: {
-    social: Social.Whatsapp,
-    webUrl: 'https://api.whatsapp.com/send?text=',
-  },
-  [Network.telegram]: {
-    social: Social.Telegram,
-    webUrl: 'https://telegram.me/share/msg?url=',
-  },
-  [Network.viber]: {
-    social: Social.Viber,
-    supportsURL: false,
-  },
-  [Network.imessage]: {
-    social: Social.Sms,
-    isNative: true,
-    supportsURL: false,
-    url: 'sms://&body=',
-  },
-  [Network.twitter]: {
-    social: Social.Twitter,
-    webUrl: 'https://twitter.com/intent/tweet?text=',
-  },
 }
